averagescore_added_to_Releasepage

+++ Releases/releases/ReleasesApp/models.py
@@ -14,8 +14,8 @@ class Release(models.Model):
     release_date = models.DateField()
     title = CharField(max_length=100)
     artist = CharField(max_length=100)
-    averagescore = models.IntegerField(validators=[MinValueValidator(0), MaxValueValidator(10)], default=-1)
-    hottestvalue = models.IntegerField(validators=[MinValueValidator(0), MaxValueValidator(10)], default=-1)
+    averagescore = models.FloatField(validators=[MinValueValidator(0), MaxValueValidator(10)], default=-1)
+    hottestvalue = models.FloatField(validators=[MinValueValidator(0), MaxValueValidator(10)], default=-1)
 
     def __str__(self):
         return self.title
@@ -24,5 +24,5 @@ class Release(models.Model):
 class ReleaseScore(models.Model):
     user = models.ForeignKey('User', on_delete=models.CASCADE)
     release = models.ForeignKey('Release', on_delete=models.CASCADE)
-    score = models.IntegerField(validators=[MinValueValidator(0), MaxValueValidator(10)])
+    score = models.FloatField(validators=[MinValueValidator(0), MaxValueValidator(10)])
